requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
djangorestframework==3.15.2
sqlparse==0.5.1
typing-extensions==4.12.2
drf-yasg==1.21.8
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-decouple==3.3
psycopg2-binary==2.9.10
dj-database-url==2.3.0
Pillow==10.4.0